requirements.txt
alembic==1.6.5
SQLAlchemy==1.4.18
flake8==3.9.2
pycodestyle==2.7.0
psycopg2==2.9.1
Flask==2.0.1
python-dotenv
flask-request-validator==4.2.0
gunicorn==20.1.0
